frontend/requirements.txt
streamlit==1.30.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.4.0
plotly==5.18.0
pandas==2.1.4
streamlit-option-menu==0.3.6
numpy==1.24.3
